Lots of changes, things seem to be working
