--- src/components/DataEntryPage/DataEntryPage.module.css
@@ -1,1 +1,0 @@
-.DataEntryPage {}
