--- src/components/Debt/Debt.module.css
@@ -1,1 +1,0 @@
-.Debt {}
