--- src/components/ResultsPage/ResultsPage.module.css
@@ -1,1 +1,0 @@
-.ResultsPage {}
