+++ src/components/ResultsPage/ResultsPage.tsx
@@ -1,27 +1,29 @@
+import { Button } from '@material-ui/core';
 import { IScenarioIoPair } from '../../interfacesAndEnums';
 import { ScenarioView } from '../ScenarioView/ScenarioView';
 import { SimulationAllResultsComparison } from '../SimulationAllResultsComparison/SimulationAllResultsComparison';
-import styles from './ResultsPage.module.css';
 
 interface ResultsPageProps
 {
-    scenarios: IScenarioIoPair[];
+  runSimulation(): void;
+  scenarios: IScenarioIoPair[];
 }
 
-export function ResultsPage(props: ResultsPageProps)
+export function ResultsPage( props: ResultsPageProps )
 {
-    return (
-        <div className={ styles.ResultsPage } >
-            { props.scenarios.length > 0 &&
-                <>
-                    <SimulationAllResultsComparison scenarios={ props.scenarios } />
-                    <table>
-                        <tr>
-                            <td><ScenarioView scenarios={ props.scenarios } /></td>
-                            <td><ScenarioView scenarios={ props.scenarios } /></td>
-                        </tr>
-                    </table>
-                </> }
-        </div>
-    );
+  return (
+    <div>
+      { props.scenarios.length > 0 &&
+        <>
+          <Button onClick={ props.runSimulation } color="primary">Run Simulation</Button>
+          <SimulationAllResultsComparison scenarios={ props.scenarios } />
+          <table>
+            <tr>
+              <td><ScenarioView scenarios={ props.scenarios } /></td>
+              <td><ScenarioView scenarios={ props.scenarios } /></td>
+            </tr>
+          </table>
+        </> }
+    </div>
+  );
 }
